Move osm to submodule
